Cosmetic changes, analytical potential function upgrade
Added **kwargs argument to the analytical potential creation function: convenient when we want the function to be dependent not only on control values and grid parameters but on other things as well, e.g. dot size, separation, etc. The **kwargs argument conveniently does not affect other parts of the code
Also, a few name changes (for better readability) and typo fixes.